refactor(landingpage): group assets
