+++ src/stories/Pages/LandingPage/sections/BackedBy.stories.tsx
@@ -1,9 +1,9 @@
 import type { Meta, StoryObj } from '@storybook/react';
 // TODO: update optimize logo files
-import AlchemyLogo from '../../../../assets/backed-by/alchemy.svg';
-import MaskLogo from '../../../../assets/backed-by/mask.svg';
-import FoothillLogo from '../../../../assets/backed-by/foothill.svg';
-import OrangedaoLogo from '../../../../assets/backed-by/orangedao.svg';
+import AlchemyLogo from '../../../../assets/LandingPage/backed-by/alchemy.svg';
+import MaskLogo from '../../../../assets/LandingPage/backed-by/mask.svg';
+import FoothillLogo from '../../../../assets/LandingPage/backed-by/foothill.svg';
+import OrangedaoLogo from '../../../../assets/LandingPage/backed-by/orangedao.svg';
 import { BackedBy } from '../../../../pages/LandingPage/sections/BackedBy';
 
 const meta = {
